Commiting the frontend main.yaml corrected the package name.

+++ ansible/terminate.sh
@@ -2,6 +2,12 @@
 
 
 TERMINATE_INSTANCE() {
+
+SPOT_INSTANCE_ID=$(aws --region us-east-1 ec2 describe-instances --filters "Name=tag:Name,Values=${COMPONENT}"|jq .Reservations[].Instances[].SpotInstanceRequestId|xargs -n1)
+if [ -z "${SPOT_INSTANCE_ID}" ]; then
+  echo "component name is wrong or empty"
+  exit 1
+fi
 aws --region us-east-1 ec2 cancel-spot-instance-requests --spot-instance-request-ids ${SPOT_INSTANCE_ID}
 INSTANCE_ID=$(aws ec2 describe-instances --filters "Name=tag:Name,Values=${COMPONENT}"|jq .Reservations[].Instances[].InstanceId|xargs -n1)
 aws --region us-east-1 ec2 terminate-instances --instance-ids ${INSTANCE_ID}
@@ -18,8 +24,4 @@ if [ "${1}" == "all"  ]; then
     TERMINATE_INSTANCE
 fi
 
-SPOT_INSTANCE_ID=$(aws --region us-east-1 ec2 describe-instances --filters "Name=tag:Name,Values=${COMPONENT}"|jq .Reservations[].Instances[].SpotInstanceRequestId|xargs -n1)
-if [ -z "${SPOT_INSTANCE_ID}" ]; then
-  echo "component name is wrong or empty"
-  exit 1
-fi
+
